Add wrapper for svg logo that sizes naturally
