+++ styles/variables.scss
@@ -27,3 +27,5 @@ $gradient-color-2: #272a92;
 $gradient-color-3: #00e7d6;
 $gradient-color-4: #0780ba;
 $gradient-speed: 0.000003;
+
+$logo-hover-transition: 0.2s;
